simplify os modal
