+++ frontend/pages/hosts/details/OSSettingsModal/OSSettingsTable/OSSettingStatusCell/OSSettingStatusCell.tsx
@@ -13,8 +13,12 @@ import {
 } from "interfaces/mdm";
 import TooltipWrapper from "components/TooltipWrapper";
 
-import { OsSettingsTableStatusValue } from "../OSSettingsTableConfig";
+import {
+  IHostMdmProfileWithAddedStatus,
+  OsSettingsTableStatusValue,
+} from "../OSSettingsTableConfig";
 import TooltipContent from "./components/Tooltip/TooltipContent";
+import generateErrorTooltip from "./errorTooltipHelpers";
 import {
   isDiskEncryptionProfile,
   LINUX_DISK_ENCRYPTION_DISPLAY_CONFIG,
@@ -34,6 +38,7 @@ interface IOSSettingStatusCellProps {
   profileName: string;
   hostPlatform?: ProfilePlatform;
   profileUUID?: string;
+  profile?: IHostMdmProfileWithAddedStatus;
 }
 
 const OSSettingStatusCell = ({
@@ -42,6 +47,7 @@ const OSSettingStatusCell = ({
   profileName = "",
   hostPlatform,
   profileUUID,
+  profile,
 }: IOSSettingStatusCellProps) => {
   let displayOption: ProfileDisplayOption = null;
   if (hostPlatform === "linux") {
@@ -65,14 +71,14 @@ const OSSettingStatusCell = ({
       case "delivered":
         if (operationType === "install") {
           displayOption = {
-            statusText: "Enforcing (pending)",
+            statusText: "Enforcing",
             iconName: "pending-outline",
             tooltip:
               "The host is running the command to apply settings or will run it when the host comes online.",
           };
         } else {
           displayOption = {
-            statusText: "Removing enforcement (pending)",
+            statusText: "Removing enforcement",
             iconName: "pending-outline",
             tooltip:
               "The host is running the command to remove settings or will run it when the host comes online.",
@@ -120,34 +126,50 @@ const OSSettingStatusCell = ({
 
   if (displayOption) {
     const { statusText, iconName, tooltip } = displayOption;
+
+    // For failed status, use the error detail as tooltip content
+    const errorTooltip = profile ? generateErrorTooltip(profile) : null;
+
+    let tipContent: React.ReactNode;
+    if (tooltip) {
+      if (status !== "action_required") {
+        tipContent = (
+          <span className="tooltip__tooltip-text">
+            <TooltipContent
+              innerContent={tooltip}
+              innerProps={{
+                isDiskEncryptionProfile: isDiskEncryptionProfile(profileName),
+              }}
+            />
+          </span>
+        );
+      } else {
+        tipContent = (
+          <span className="tooltip__tooltip-text">
+            <TooltipContent
+              innerContent={tooltip}
+              innerProps={{ isDeviceUser, profileName }}
+            />
+          </span>
+        );
+      }
+    } else if (errorTooltip) {
+      tipContent = (
+        <span className="tooltip__tooltip-text">{errorTooltip}</span>
+      );
+    }
+
     return (
       <span className={baseClass}>
         <Icon name={iconName} />
-        {tooltip ? (
+        {tipContent ? (
           <TooltipWrapper
-            tipContent={
-              <span className="tooltip__tooltip-text">
-                {status !== "action_required" ? (
-                  <TooltipContent
-                    innerContent={tooltip}
-                    innerProps={{
-                      isDiskEncryptionProfile: isDiskEncryptionProfile(
-                        profileName
-                      ),
-                    }}
-                  />
-                ) : (
-                  <TooltipContent
-                    innerContent={tooltip}
-                    innerProps={{ isDeviceUser, profileName }}
-                  />
-                )}
-              </span>
-            }
+            tipContent={tipContent}
             position="top"
             underline={false}
             showArrow
             tipOffset={8}
+            clickable
           >
             <span className={`${baseClass}__status-text`}>{statusText}</span>
           </TooltipWrapper>
