--- frontend/pages/hosts/details/OSSettingsModal/OSSettingsTable/OSSettingsErrorCell/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./OSSettingsErrorCell";
