+++ frontend/pages/hosts/details/OSSettingsModal/OSSettingsTable/OSSettingsNameCell/_styles.scss
@@ -7,8 +7,4 @@
   }
 }
 
-// need to do some overrides here to get the name text content to be the right size
-.data-table-block .data-table tbody td .os-settings-name-cell__name-tooltip {
-  max-width: 142px;
-  min-width: auto;
-}
+
