+++ frontend/pages/hosts/details/OSSettingsModal/OSSettingsTable/OSSettingsResendCell/_styles.scss
@@ -1,20 +1,8 @@
-.os-settings-error-cell {
+.os-settings-resend-cell {
   display: flex;
-  justify-content: space-between;
+  justify-content: flex-end;
   align-items: center;
   gap: $pad-small;
-
-  &__failed-message {
-    // allows the message to shrink and not push the
-    // resend button outside of the table cell
-    min-width: 0;
-
-    .data-table__tooltip-truncated-text--cell {
-      // for some reason this is need to vertically align the text and
-      // the resend button
-      display: block;
-    }
-  }
 
   &__resend-button {
     width: 106px;
