+++ frontend/pages/hosts/details/OSSettingsModal/OSSettingsTable/OSSettingsResendCell/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./OSSettingsResendCell";
